Update README: fix download links (is.gd -> kutt.it)
As after years of operation, is.gd currently seems to be offline.

The list of free url shorteners that allow custom URLs is quite short, but kutt.it was the first alternative I could find that works just as well. Personally, I find custom URLs with memorizable words to be a lot easier to work with, than trying to type in some random characters on a TV remote.